wallet: extract RenameAccount shared workflow
